Restructure stop area models
