+++ lib/models/stop_area_processing/stop.dart
